refactor: rename history events module
