+++ src/historyView/index.html
@@ -19,7 +19,7 @@
       {
         "imports": {
           "./modules/domHandlers.js": "${domHandlersUri}",
-          "./modules/uiManagers/historyEvents.js": "${historyEventsUri}",
+          "./modules/uiManagers/Events.js": "${eventsUri}",
           "./modules/uiManagers/historyRenderer.js": "${historyRendererUri}",
           "./modules/historyViewState.js": "${historyViewStateUri}",
           "./modules/uiManagers/searchManager.js": "${searchManagerUri}",
